requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
numpy==1.18.4
pandas==1.0.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyyml==0.0.2
pyzipcode==2.0.0
requests==2.23.0
six==1.14.0
tabulate==0.8.7
urllib3==1.25.9
